+++ packages/anatomy/src/fieldset/index.ts
@@ -1,0 +1,3 @@
+import { createAnatomy } from '@zag-js/anatomy'
+
+export const anatomy = createAnatomy('fieldset').parts('root', 'errorText', 'helperText', 'legend')