+++ packages/react/src/components/fieldset/examples/with-field.tsx
@@ -1,0 +1,17 @@
+import { Field, Fieldset } from '../..'
+
+export const WithField = (props: Fieldset.RootProps) => {
+  return (
+    <Fieldset.Root {...props}>
+      <Fieldset.Legend>Legend</Fieldset.Legend>
+      <Fieldset.HelperText>Fieldset Helper Text</Fieldset.HelperText>
+      <Fieldset.ErrorText>Fieldset Error Text</Fieldset.ErrorText>
+      <Field.Root>
+        <Field.Label>Label</Field.Label>
+        <Field.Input />
+        <Field.HelperText>Field Helper Text</Field.HelperText>
+        <Field.ErrorText>Field Error Text</Field.ErrorText>
+      </Field.Root>
+    </Fieldset.Root>
+  )
+}