+++ packages/react/src/components/fieldset/use-fieldset-context.ts
@@ -1,0 +1,11 @@
+import { createContext } from '../../utils/create-context'
+import type { UseFieldsetReturn } from './use-fieldset'
+
+export interface UseFieldsetContext extends UseFieldsetReturn {}
+
+export const [FieldsetProvider, useFieldsetContext] = createContext<UseFieldsetContext>({
+  name: 'FieldsetContext',
+  hookName: 'useFieldsetContext',
+  providerName: '<FieldsetProvider />',
+  strict: false,
+})